requirements.txt
# 视频处理核心依赖
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# 视频处理
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# 面部检测和分析
dlib>=19.24.0
face-recognition>=1.3.0
mediapipe>=0.10.0

# 数据处理
pandas>=2.0.0
tqdm>=4.65.0

# 日志和配置
loguru>=0.7.0
pyyaml>=6.0
click>=8.1.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
